--- index.d.ts
@@ -1,5 +1,0 @@
-export * from './src/pageObject';
-export * from '@qavajs/playwright-runner-adapter';
-export * from './src/QavajsPlaywrightWorld';
-export * from './src/QavajsPlaywrightElectronWorld';
-export type * from './src/types';
