+++ package.json
@@ -2,7 +2,8 @@
   "name": "@qavajs/playwright",
   "version": "3.5.0",
   "description": "steps to interact with playwright",
-  "main": "./index.js",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
   "scripts": {
     "build": "tsc",
     "start:server": "ts-node test-e2e/support/server.ts",
@@ -34,7 +35,7 @@
   },
   "dependencies": {
     "@playwright/test": "^1.55.1",
-    "@qavajs/playwright-runner-adapter": "^1.5.1",
+    "@qavajs/playwright-runner-adapter": "^1.6.0",
     "@qavajs/memory": "^1.10.2"
   }
 }
