+++ src/QavajsPlaywrightElectronWorld.ts
@@ -23,8 +23,8 @@ const electron = test.extend<ElectronFixture>({
 export class QavajsPlaywrightElectronWorld extends QavajsPlaywrightWorld {
     test = electron;
     electron!: ElectronApplication;
-    context!: BrowserContext;
-    page!: Page;
+    declare context: BrowserContext;
+    declare page: Page;
     constructor(options: any) {
         super(options);
     }
