+++ test-e2e/playwright.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig, devices } from '@playwright/test';
-import { defineCucumber } from '../index';
+import { defineCucumber } from '../lib';
 
 /**
  * Read environment variables from file.
